Update bzl gen for python to look at .pyi files when computing deps
Summary:
Include depedencies from .pyi files in the computed dependencies for a
dbx_py_library. This mostly impacts protobufs, which frequently refer to
`dropbox.model.types` and `dropbox.model.dbtypes`.

Properly tracking this dependency is important for bazeled mypy, since
without the dependency, it will produce invalid caches that are
invalidated when checking targets that themselves depend on the
missing dependency, causing lots of rechecking.

I'm putting the diff up without the regenerated BUILD files to cut
down on noise.

A few other minor knock-on changes were needed to deal with the new
dependencies.

Test Plan:
After regenerating everything, run a mypy bazel test and verifying we
aren't rechecking all the protos because of added deps.

GitOrigin-RevId: bea09cc9d6a3a0a18449aa27e0442111c7267352