Batch BUILD merging in bzl gen
Summary:
Shelling out is expensive. I played with doing it with asyncio.subprocess but it didn't
help much (process creation is too expensive). And batching is simpler to implement anyway.
It doesn't seem like we need to bother parallelizing on the go side (if we do, we need
to modify the python side a bit to not rely on ordering)

The maximum batch size is simply so that we don't trip the os limit for maximum # of args
passed to the process, although I'm not sure that this is a problem in practice.

GitOrigin-RevId: d198b7591e73ce07ae2cf30d6d794a1e19ed2b24