Make the steps from the README file work
When used as an external dependency dbx_build_tools and its tools
are unpacked in an "external" subdirectory. Following the steps in
the README file, adding dbx_build_tools as an http_archive, it is
unpacked in this way. This causes the itest-* tooling and others
to not work as the scripts are looking in the wrong place for
their runfiles.

Since dbx_build_tools is not always used as an external dependency,
some logic to figure out where the runfiles directory can be found
is required.