[windows] Create a Windows-worthy environment for vpip
Summary:
Windows has a somewhat different set of environment
variables that it needs to have for vpip to work,
so this is a first stab at the flags that we need.

A lot of the environment analysis is pretty similar
to the MacOS/Linux versions, but many of the flags
we set are different.

GitOrigin-RevId: 4992c194fed337d4dfd3b8991905ca10c5f834e6