[windows] Add a runfiles wrapper for Windows
Summary:
This offers an attempt to implement the runfiles
wrapper in using a batch file, assuming that the
system has Python 3.8 installed. The basic idea
is to reimplement the current runfiles detection
process.

System Python is used to resolve symlinks and
get the real path of the runfiles directory, which
is why we necessitate Python 3.8. `realpath` is not
supported on Windows pre-3.8.

GitOrigin-RevId: fa2bd445cd44cef0f0db1eb83b5cf87a06e94bb3