[py] Allow check_conflicts to take an argument file
Summary:
On Windows, `check_conflicts` can (and will) fail due to excessively
long command lines -- I have hit this when trying to build the Cython
wrapper. In order to avoid this, we force Bazel to write the arguments
to an argument file (Bazel is not smart enough to do this on its own),
and restructure the `check_conflicts` script to appropriately handle
it.

GitOrigin-RevId: 42a59bfd0a2eefdf34e0cf2a1126dd975bb8be40