[py] Make cffi more Windows-friendly
Summary:
On Windows, we need the default shell environment
so that we can find system Python.

Pretty much the same reason as all the other places
where we added `use_default_shell_env`.

On Windows, we also need to pass in the system
libraries that it needs (e.g. `shcore`, `user32`), so this
adds a `libraries` field for that. We can achieve the
same thing by adding `libname.lib` to `linkopts`, but
that seems ugly, since on Windows the libraries
to link are arguments and not really options.

Windows also generally wants Unicode to be enabled
with FFI.

Test Plan: - Things ought to still build.

GitOrigin-RevId: 7e58fc64106015b13e97e3827d6018f2b7191523