[vpip] Allow MSVC linking to accept arbitrary inputs
Summary:
On Windows, we occasionally want to link to objects that
aren't "libraries". For example, we want to
link against a `version.res` file for CFFI extensions that
specify the DLL file version. The Bazel way to do this is
to pass the `.res` file as a linker "flag" (actually an
argument). See https://github.com/bazelbuild/bazel/issues/8273
for the details.

Since on Windows we specify linker inputs as simple arguments,
we just pass objects we want to link against in the `LINK`
environment variable, instead of the convoluted `--libraries`
and `--library-dirs` solution we had before.

We also relax the linker flag restrictions on Windows, which
will let us pass the object paths as-is to `vpip`.

GitOrigin-RevId: 1e3568fcb26ab3a4fa833d8f971881616063f0f3