[gazel] Handle build folders gracefully
Summary:
If you happen to have a folder named `BUILD` (on Linux)
or maybe a folder named `build` on a case-insensitive
system like Windows/OSX, `gazel` will choke trying to
write to it.

To prevent this, we defer writing to the output file
until the end, when we've properly determined the
right output path to use.

To make sure that the name conflict check is still
accurate, we check if `BUILD`/`BUILD.bazel` are files,
and if they are, we just delete them instead of
overwriting them. The added benefit to deleting is that
on case-insensitive systems, we'll end up "correcting"
the case -- i.e. deleting a `build` file and writing
a new `BUILD` file.

GitOrigin-RevId: 8e78fabaefe8087bb07e1e79511235c9be3facc9