[py] Let setuptools invoke cffi
Summary:
Instead of having FFI emit C code and then passing
that C code to setuptools (which drops some of the
FFI internal state, like that in `set_unicode`),
just have setuptools invoke FFI and handle it all
in one fell swoop.

This resolves an issue on Windows `dbx_cffi_module`
can silently and incorrectly compile Unicode-specific
C code, which is a fairly scary prospect. This happens
because FFI stores the `set_unicode` state internally
and expects to pass the `UNICODE`/`_UNICODE` define
macros to `distutils` directly.

Note that this change now makes it so that CFFI modules
must have the source and cdef in the same directory
as the BUILD file (either the source file or a generated
file). This is because we have removed the intermediate
"generate a C file" step, and how the `sdist` tar is created
in our pip rules.

GitOrigin-RevId: 2ab699ff8a3b163c8f31f5733a073b4f62c5c3e1