Fix a bug in `vpip.py` where the local site can interfere with installing tooling wheels
Summary:
This fixes a slight bug I experienced on my Linux VM: I had a more recent version of `pip` and `setuptools` in my user site (i.e. `~/.local`) and `vpip` would try to uninstall those when trying to install its vendored copies of `pip`, `setuptools` and `wheel`: this is (1) bad because Bazel shouldn't care about this and (2) can't work because sandboxing prevents the operation from succeeding.

The typical fix for this is to prevent the user site from being considered, so adding `PYTHONNOUSERSITE` fixes it. I believe this is aligned with what we're trying to achieve here, which is isolation from the environment.

Test Plan: Tested locally on Ubuntu 20.04, works great.

GitOrigin-RevId: 2031efcf4c8d2b4712931b6b69b3dd6d5d984c5e