[svcinit+bzl] Add is_critical attribute
Summary:
This diff introduces the notion of "critical services". If a bazel service is marked critical, any
tests that depend on the service will be marked failed if the service is not healthy when the tests
complete, regardless if the tests pass.

The feature is implemented as follows:

1. Add a flag `--svc.critical-services` in `svcinit`. It accepts a comma separated list of service
   names that `svcinit` should check their health when the test run complete.
2. Update `svcinit` to check these "critical services" health after test run complete.
3. Add an attribute `is_critical` to the bazel rule `dbx_dropbox_service_daemon`. Users can mark the
   service as critical by setting the attribute. Its default value is `False`.
4. Update various rules in `svc.bzl` to plumb through the attribute.
5. Update `_services_bin_impl` to collect all critical services in the dependent services and pass
   them into `svcinit` via the added flag in (1).

NOTE: I admit the notion of "critical services", and the attribute "is_critical" are not very
descriptive. I am open to suggestion to better names. Maybe we can name the attribute as
`fail_test_on_crash`?

GitOrigin-RevId: 474c664c80ba65975c4f42476b1130be3301326d