[deps] Make OpenSSL build on macos and update
We change up the openssl build generation to be able to generate builds
for other platforms and systems.

The initial added platforms are for macos. In the process of doing this
the build generation tooling has changed to be a python script that
understands all of the select permutations needed to generate the
stanzas needed for more than one platform.

Alongside this the perl script used to extract out the build details has
been changed. Rather than generating the build rules directly it now
exports the build information needed in the form of JSON outputs.

This has been tested on x86_64 for macos catalina and linux.

Right now there are probably some unixisms left in the code which will
limit its ability to build for windows.