[ML Platform] Add dynamic library support to dbx_py_binary, use it to load CUDA
Summary:
CUDA is not available in source form. We keep the binary shared libraries in Magic Mirror, and so far have been bringing it into our Python binaries by including it in a `data=` attr, then adding `LD_LIBRARY_PATH` when executing them. It's a hack. This diff updates `dbx_py_binary` (and, transitively, `dbx_py_pytest_test`) to add a `dynamic_libraries` attribute, which takes a `filegroup` with the set of shared objects. They're added to the existing machinery for supporting `.so`'s that were built by bazel. Note that we can't currently write tests that depend on CUDA because we don't have GPU support in Changes.

In addition, the `@cuda//cuda:cuda-lib` target is very large. This is partly because the symlinked shared objects are all expanded into whole files (so we get multiple copies of each), and partly because libraries that are not needed are still included.

This diff creates a new target, `@cuda//cuda:cuda-lib-tf` that includes only the libraries needed by TensorFlow, under names that TF's dso loader will use to search for them (tf doesn't use ld to link CUDA). This substantially reduces the size of TF GPU binaries (from 16G or more to ~3.3G)

Also removes a workaround for dealing with some old bazel bugs, that doesn't seem to be necessary any longer. It doesn't change the binary size, but should make working with CUDA in devbox a little faster.

Test Plan: TF with GPU should work without weird hackery

GitOrigin-RevId: 7d9d4f6492f32d6b0ae312b537c3535eb60b03eb