moving godep.bzl to be under build_tools
Summary:
This change is inspired by a breakage in Vagrant rsync:
https://sourcegraph.pp.dropbox.com/server/-/blob/desktop/Vagrantfile?L105-122&subtree=true
The previously added godep.bzl was located in the root of rSERVER.
This is unnecessary and may cause hard coded logic such as this one to break.
Moving it to be under build_tools makes more sense.

Test Plan: Same testing procedure as https://phabricator.dropboxer.net/D826936

GitOrigin-RevId: 9e74d4f9754cda5e7c2c3167defe52b9ee30cab9