[py] Remove linker flag if library is directly passed
Summary:
If we pass in a static library to link against, the
`ld` wrapper doesn't remove the corresponding `-l`
flag if it exists, which causes problems outside of
server code (when ignore missing static libraries is
disabled).

This diff makes it so that the linker doesn't look for
a library if it's passed in directly to the wrapper.

This will be useful for getting `whoseisit` to build
for M1.

Test Plan:
- Built `pyyaml` (which relies on `libyaml`) and observed
  that the C extension was built appropriately on macOS.
  It was failing (and silently reverting to pure-Python)
  beforehand.

GitOrigin-RevId: 1fac1714b9924f0b3389a08c9d55f2530ce32598