[py] Fix a depset issue with dynamic_library handling
Summary:
`depset` is supposed to take a sequence of `depset`s in the
transitive field. We can't use `to_list()` into `transitive`
it won't match types unless the array is empty.

Since `dynamic_libraries_trans` is a `depset` already, the
simplest solution is to just wrap it in an array and pass it
along.

This fixes issues with desktop targets (which also use dynamic
libraries, but not with the `dynamic_libraries` field) where
it was improperly unwrapping a `depset` of files.

Test Plan:
- Built `tensorflow_gpu` and it didn't error out.
- Patched it into desktop code and verified that the build
  process Python targets with dynamic libraries didn't error out.

GitOrigin-RevId: 1931b57a45e03dc805519907d371e64296bc348e