Remove tensorflow-addons
Summary:
This package is

1) Unused
2) Using unsupported build tools for binary package
3) Of something that does not have a published cpython-39 wheel

The options are to upgrade, build from source, or to delete it. Since it is unused
I am opting to delete it. It was added in Nov 2021 and not used yet. If the author
of that diff cares to use it, they can add it using supported build mechanisms.

GitOrigin-RevId: f406c112e23e2669a94a03816dcd3415f97fc7b4