Add go 1.17 support
Summary:
Go 1.18's expected release date is Feb. 2022, so Go 1.16 is reaching its [[ https://endoflife.date/go | end of life ]], and we won't be receiving updates.

This means it's necessary to start moving things over to a new version of Go, hence why we're adding Go 1.17 support and getting the ball rolling on updating some of our Go binaries.

Test Plan: Update random stuff to Go 1.17, see if anything breaks. If so, revert.

GitOrigin-RevId: b071c3f9ee8a5c6eb1c2da23e026e40f91c724a7