Remove Go 1.17 support, add Go 1.18 support
Summary:
This is redoing D836958. We need to upgrade Go because 1.18 is being released soon,
so 1.16 is reaching its end of life. We've decided to update directly to 1.18. Since 1.18
isn't actually released yet, we'll use a pre-release version, then when 1.18 is actually released,
we'll swap that out for the stable version. We won't be changing the default version to 1.18
until a few weeks after the stable 1.18 release.

I'll enable Go 1.18 tests in a separate diff; last time athena reverted my Go 1.17 Tests diff
because it was causing infra failures due to increased load on changes.

Test Plan: N/A

GitOrigin-RevId: 06258beb22010014fd67d058035578898c26b9a2