Remove //thirdparty/typeshed
Summary:
This folds //thirdparty/typeshed into //thirdparty/mypy and removes the vendored files.
Instead, we declare all the typeshed related stuff in @mypy// and change all the references
in the codebase to refer to the new locations for bazel mypy.

This is more or less the last thing needed to open source the bazel mypy rules properly
(which is required if we ever want to compile something like bzl which is open source)
and should make updating a pin less of a pain.

Test Plan: bzl test //dropbox/mypy/test_mypyc/...

GitOrigin-RevId: 0b213f7958caaf2b9008d131c4eb6574e522b9e2