Erase drte from the knowledge of mypy
Summary:
This fixes a mistake I made when I moved the mypy build rules to use toolchains -
I made it parameterized by drte. There isn't a real need to do this - the python
rules should be taking care of this for us. Unfortunately for some reason the
interpreter doesn't include the cc_library form of the headers, which is what we
want, so I added cc_headers to differentiate that from headers.

The dbx_mypy_runtime rule does all the hard work - its like a cc_library, but
it injects the py headers from whatever python toolchain has been chosen.

Subsequently I renamed all the mypy toolchains to remove the reference to drte.
This will allow us to open source //thirdparty/mypy as it no longer has a direct
dependency (at last) on the python headers for the mypy runtime (and should be possible
to use the non-prebuilt toolchains)

Now that drte is not involved, I am not sure toolchains are really necessary for us
here. However, I think it still makes things a little cleaner.

GitOrigin-RevId: 16b50cd7adcd9b7642f0dd26bc6c392c735533b0