Prepare //thirdparty/mypy for public release.
Summary:
This moves the workspace declaration to use http_archive so its publicly accessible
and removes our custom patch in favor of just adding a //pip/types-typed-ast dependency.

After this it should be in a clean enough state to release the mypy binary itself (although
we still need to clean up the build rules to not depend on our internal plugins before
they will function).

Test Plan: mypy continues to build

GitOrigin-RevId: 7421473e533fe8952d9d6e083fcce27702669cc7