Quick hack to give mypy correct version under bazel
Summary:
I broke the mypy cache warming because the new caches don't have proper mypy version
(just 9.20+dev). It will take time to make a proper fix like accepting an environment
variable in upstream setup.py but we need the mypy warming to work now, I am just patching
it in.

We can probably remove this patch on the next pin upgrade we do by upstreaming a change
to setup.py that fixes the issue.

GitOrigin-RevId: 30ea215370fcbbc16627bf3a410e79c5fffffb8a