[bzl] pkg rules now allow relative data labels
Summary:
Currently if you try to have relative labels as a `data` dependency
of a `dbx_pkg_deb` rule, `bzl` interprets it as a target in the
repo root. This is fantastically wrong.

This fixes the issue by properly resolving the full target name,
leveraging the fact that `BazelTarget` mostly does this for us,
and that `bazelpkg` always changes its directory to the workspace
directory when it does its building.

Test Plan:
- Made a `dbx_pkg_deb` with a relative data dependency, verified
  that it built properly.

GitOrigin-RevId: fd4878a00caa0dc799eef67c69ec4fd03a0e0ed4