[Bzl Gen] Allow dbx_go_test to use the go_versions attribute
Summary:
This diff allows the usage of `go_versions` for `dbx_go_test` rules.
The reason is that some tests have deps that are incompatible with
Go 1.18, but if we upgrade them they become incompatible with Go
1.16. Therefore, we need to add the ability to have the tests
running under one version of Go.

Test Plan:
Run `BZL_DONT_USE_SQPKG=1 bzl gen //foo` and observe that
the generator fails if I set `go_versions` for a dbx_go_binary, but
succeeds if I set `go_versions` for a dbx_go_test.

GitOrigin-RevId: e77ce8cf85e3154530edcd1f55979742745bb2a9