Create a test container image that bakes in drte
This diff creates a new test container image that bakes in a copy of
drte-v5, plus its dependencies.

I spun up a copy of the container using docker in my devbox and
confirmed that the binaries I copied inside work as expected.

Next steps are to try pushing the modified deployment and confirm
both versions of the ping_server start up and pass health checks.

GitOrigin-RevId: 200aebb102a4dcfb262bc963cb023c9a088c584a