Fix rust deps of piplib targets.
Adding a rust_library as a dep of a piplib target results in a skylark
type error because the format of rust_lib changed in the past.

The last time this code was used was ~2020 so maybe it would be better
to remove the support entirely. But this change seems strictly better
than the status quo.

add a dep, check that the error message goes away

GitOrigin-RevId: 297110a3804e75b0d11c2b3a874e216476e3d64a