[sqfs] Tolerate short dest duplicates if they have the same hash
Upstream bazel go rules use configuration transitions for go tags, race detector, etc, which outputs files into a separate target location for each such config.
This may lead to many different variations of the same file in the manifest, but hopefully all of them have the same actual content.

This change relaxes constraint for the manifest to allow duplicates, but it validates that all of them have exactly the same hash for correctness.

GitOrigin-RevId: e641ba622c755d4c8bebacacb0ca577af1363d2f